enhanced ui and fixed bug

+++ app/src/main/java/com/example/wellmate/data/HealthScore.kt
@@ -5,19 +5,13 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "health_scores")
 data class HealthScore(
-    @PrimaryKey(autoGenerate = true) val id: Int = 0,
-
+    @PrimaryKey(autoGenerate = true)
+    val id: Long = 0L,
     val userId: String,
-
     val timestamp: Long,
-
-    val score: Int,
-
+    val score: Float,               // FLOAT to store decimal model outputs
     val category: String,
-
     val reason: String,
-
     val confidence: Float,
-
     val recommendationJson: String
 )
